Use a running average and peak detection.
Calculate a running average and use peak detection in an attempt to
eliminate false positives.  Utilizes a noise band that minimally affects
the running average.  Values outside of the noise band have a greater
effect on the running average.  Finally, values outside of 2x the noise
band are completely ignored, as they're assumed to be anomalous.

Also:

* Reduces the number of pads (2 broke).
* Adds a trigger limit so that at least 2 pads must be active for
  presence detection.
* Longer calibration on startup w/ a greater initial delay before
  performing calibration.